Added support to display an MOTD on connect (#121)
* Moved configuration-related stuff to a new configuration module

This will allow sessions to be able to get and set config values without having to know about the whole add-on.

* Display an MOTD from the server if provided

If the MOTD is the same one that the server sent the last time, it will be ignored unless the optional field force_display = True
